pyproject.toml
[project]
name = "tech-friday-genai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.11.0",
    "google-genai>=1.21.1",
    "sentence-transformers>=4.1.0",
    "serpapi>=0.1.5",
]
